backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.6

# Error Tracking
sentry-sdk[fastapi]==1.40.0

# GitHub
PyGithub==2.1.1
PyJWT==2.8.0
cryptography==42.0.0

# AI/LLM
openai==1.10.0
# anthropic==0.8.1  # Optional - uncomment if using Anthropic

# Linting/Analysis
pylint==3.0.3
flake8==7.0.0
bandit==1.7.6

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
